Added Environment based specs

+++ backend/internal/services/auth_service.go
@@ -37,15 +37,17 @@ func NewAuthService(userRepo repositories.UserRepository, jwtService utils.JWTSe
 }
 
 func (s *authService) Signup(req *dtos.SignupRequest) (*dtos.AuthResponse, uint, error) {
-	// Check if user exists
-
 	if req.Username == config.Env.AdminUser {
 		return nil, http.StatusBadRequest, errors.New("username already exists")
 	}
 
-	validUserSignupSecret := s.userRepo.ValidateUserSignupSecret(req.UserSignupSecret)
-	if !validUserSignupSecret {
-		return nil, http.StatusUnauthorized, errors.New("invalid user signup secret")
+	if config.Env.Environment == "DEVELOPMENT" {
+		log.Println("Development mode, skipping user signup secret validation")
+	} else {
+		validUserSignupSecret := s.userRepo.ValidateUserSignupSecret(req.UserSignupSecret)
+		if !validUserSignupSecret {
+			return nil, http.StatusUnauthorized, errors.New("invalid user signup secret")
+		}
 	}
 	existingUser, err := s.userRepo.FindByUsername(req.Username)
 	if err != nil {
@@ -92,9 +94,13 @@ func (s *authService) Signup(req *dtos.SignupRequest) (*dtos.AuthResponse, uint,
 	}
 
 	go func() {
-		err := s.userRepo.DeleteUserSignupSecret(req.UserSignupSecret)
-		if err != nil {
-			log.Println("failed to delete user signup secret:" + err.Error())
+		if config.Env.Environment == "DEVELOPMENT" {
+			log.Println("Development mode, skipping user signup secret deletion")
+		} else {
+			err := s.userRepo.DeleteUserSignupSecret(req.UserSignupSecret)
+			if err != nil {
+				log.Println("failed to delete user signup secret:" + err.Error())
+			}
 		}
 	}()
 
@@ -179,9 +185,6 @@ func (s *authService) Login(req *dtos.LoginRequest) (*dtos.AuthResponse, uint, e
 }
 
 func (s *authService) GenerateUserSignupSecret(req *dtos.UserSignupSecretRequest) (*models.UserSignupSecret, uint, error) {
-	if req.Username != config.Env.AdminUser || req.Password != config.Env.AdminPassword {
-		return nil, http.StatusUnauthorized, errors.New("invalid credentials for the admin")
-	}
 
 	secret := utils.GenerateSecret()
